patches: Remove outdated patch
The patch was:
* Replacing the lua `main` symbol with `lua_main`.
* Removing checking that the given argument was a script, instead
  checking if there is an initrd filesystem present and trying to
  execute it.

The `main` symbol is already replaced in the preprocessing stage by
adding `-Dmain=lua_main` in the `Makefile.uk` flags, so renaming it is
not needed.

Trying to execute the initrd without properly extracting it will
obviously fail, since `lua` will not be able to read the script properly.
This will result in an `malformed number near '0707010..'` error.

The check for initrd is not needed at all, since the fs driver is
transparent for the application, and the script can be accessed as an
usual file, regardless of the fs driver (initrd, 9pfs, etc.).

Signed-off-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Reviewed-by: Eduard Vintilă <eduard.vintila47@gmail.com>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #8